notification_service/requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
mongoengine==0.29.1
pika==1.3.2
drf-yasg==1.21.11
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
djangorestframework-simplejwt==5.3.1

# OpenTelemetry
opentelemetry-distro==0.45b0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-django==0.45b0
# opentelemetry-instrumentation-djangorestframework # Not supported on Python 3.13 yet
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-pymongo==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-sdk==1.24.0

# Logging
python-json-logger==2.0.7

# Metrics
django-prometheus==2.3.1